[OpenWeatherMap] Split out common code to a base class
